requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dnspython==1.16.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.6
numpy==1.19.3
pandas==1.1.5
pylint==2.6.0
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
